Refactor: Extract `UserInfo` type to dedicated file and update imports
