+++ src/types/user.ts
@@ -1,0 +1,7 @@
+export interface UserInfo {
+    sub: string;
+    email: string;
+    name?: string;
+    preferred_username?: string;
+}
+
